Also try all keys during `PacSignature` verification
With the previous patch, we still have the issue that the signature
verification process uses the old way of selecting the right key from
a given key table. This means that the patch just attempting every key
during the decryption process might not come into effect, because the
initial signature check already fails (because it still only considers
one key, which might be the wrong one).

I've mainly copied over the new logic from `DecryptedKrbApReq` and only
adapted it slightly for the signature verification.